fix: 테이블 관련 타입 분리 및 index로 key값 사용할 수 있도록 허용
